Implemented custom hook to fetch prdoducts by id
